Rename residue to residue_definitions
