+++ nanofold/common/residue_definitions.py
